fix: V10's flag column must be INT, and prove appliances round-trip (KAN-23)
The IT suite caught a latent startup bug. V10 declared ate_fast_food_last_turn as
TINYINT while SaveEntity maps it as an int-as-boolean, exactly like won and rent, so
Hibernate's ddl-auto=validate rejected the schema: "found [tinyint], but expecting
[integer]". Every Testcontainers IT failed to build its context - the app would not
have started against a migrated database.

Tasks 6-11 all ran mvnw package (unit tests, no DB), which cannot see this; only
verify can. Both halves of the mismatch came from the plan's own code. V10 has not
been applied to any real database - only throwaway test containers - so correcting
the migration in place is safe rather than superseding it.

Also extends the ports adapter IT to round-trip the new state through real MySQL:
the fast-food flag, and the appliance join table. The second update asserts that
update() replaces the join-table rows rather than accumulating them, which nothing
covered before.

+++ amiss-core/src/main/resources/db/migration/V10__food_storage_and_appliances.sql
@@ -3,7 +3,10 @@
 -- never banked); appliance ownership (Fridge/Freezer here; Computer/Books
 -- follow in a later PR with zero schema change) is a generic join table, the
 -- same @ElementCollection shape as tbljob_degrees.
-ALTER TABLE tblsave ADD COLUMN ate_fast_food_last_turn TINYINT NOT NULL DEFAULT 0 AFTER eat;
+-- INT, not TINYINT: SaveEntity maps this as an int-as-boolean (0/1) exactly like `won`
+-- and `rent`, and Hibernate's ddl-auto=validate rejects a TINYINT column behind an int
+-- field ("found [tinyint], but expecting [integer]") — the app would not start.
+ALTER TABLE tblsave ADD COLUMN ate_fast_food_last_turn INT NOT NULL DEFAULT 0 AFTER eat;
 
 CREATE TABLE tblsave_appliance (
     save_id   BIGINT      NOT NULL,
